add autofill command
