+++ manifests/manifest-firefox.json
@@ -34,6 +34,9 @@
         "_execute_browser_action": {},
         "scan-qr": {
             "description": "Scan a QR code"
+        },
+        "autofill": {
+            "description": "Autofill the matched code"
         }
     },
     "options_ui": {
