+++ src/store/Accounts.ts
@@ -3,6 +3,7 @@ import { Encryption } from "../models/encryption";
 import * as CryptoJS from "crypto-js";
 import { OTPType, OTPAlgorithm } from "../models/otp";
 import { ActionContext } from "vuex";
+import { getSiteName, getMatchedEntriesHash } from "../utils";
 
 export class Accounts implements Module {
   async getModule() {
@@ -22,7 +23,7 @@ export class Accounts implements Module {
         sectorOffset: 0, // Offset in seconds for animations
         second: 0, // Offset in seconds for math
         filter: true,
-        siteName: await this.getSiteName(),
+        siteName: await getSiteName(),
         showSearch: false,
         exportData: await EntryStorage.getExport(entries),
         exportEncData: await EntryStorage.getExport(entries, true),
@@ -41,7 +42,7 @@ export class Accounts implements Module {
           );
         },
         matchedEntries: (state: AccountsState) => {
-          return this.matchedEntries(state.siteName, state.entries);
+          return getMatchedEntriesHash(state.siteName, state.entries);
         },
         currentlyEncrypted(state: AccountsState) {
           for (const entry of state.entries) {
@@ -513,82 +514,6 @@ export class Accounts implements Module {
     };
   }
 
-  private async getSiteName() {
-    return new Promise((resolve: (value: Array<string | null>) => void) => {
-      chrome.tabs.query({ active: true, lastFocusedWindow: true }, (tabs) => {
-        const tab = tabs[0];
-        const query = new URLSearchParams(
-          document.location.search.substring(1)
-        );
-
-        let title: string | null;
-        let url: string | null;
-        const titleFromQuery = query.get("title");
-        const urlFromQuery = query.get("url");
-
-        if (titleFromQuery && urlFromQuery) {
-          title = decodeURIComponent(titleFromQuery);
-          url = decodeURIComponent(urlFromQuery);
-        } else {
-          if (!tab) {
-            return resolve([null, null]);
-          }
-
-          title = tab.title?.replace(/[^a-z0-9]/gi, "").toLowerCase() ?? null;
-          url = tab.url ?? null;
-        }
-
-        if (!url) {
-          return resolve([title, null]);
-        }
-
-        const urlParser = new URL(url);
-        const hostname = urlParser.hostname; // it's always lower case
-
-        // try to parse name from hostname
-        // i.e. hostname is www.example.com
-        // name should be example
-        let nameFromDomain = "";
-
-        // ip address
-        if (/^\d+\.\d+\.\d+\.\d+$/.test(hostname)) {
-          nameFromDomain = hostname;
-        }
-
-        // local network
-        if (hostname.indexOf(".") === -1) {
-          nameFromDomain = hostname;
-        }
-
-        const hostLevelUnits = hostname.split(".");
-
-        if (hostLevelUnits.length === 2) {
-          nameFromDomain = hostLevelUnits[0];
-        }
-
-        // www.example.com
-        // example.com.cn
-        if (hostLevelUnits.length > 2) {
-          // example.com.cn
-          if (
-            ["com", "net", "org", "edu", "gov", "co"].indexOf(
-              hostLevelUnits[hostLevelUnits.length - 2]
-            ) !== -1
-          ) {
-            nameFromDomain = hostLevelUnits[hostLevelUnits.length - 3];
-          } else {
-            // www.example.com
-            nameFromDomain = hostLevelUnits[hostLevelUnits.length - 2];
-          }
-        }
-
-        nameFromDomain = nameFromDomain.replace(/-/g, "").toLowerCase();
-
-        return resolve([title, nameFromDomain, hostname]);
-      });
-    });
-  }
-
   private getCachedPassphrase() {
     return new Promise((resolve: (value: string) => void) => {
       chrome.runtime.sendMessage(
@@ -603,63 +528,5 @@ export class Accounts implements Module {
   private async getEntries() {
     const otpEntries = await EntryStorage.get();
     return otpEntries;
-  }
-
-  private matchedEntries(
-    siteName: Array<string | null>,
-    entries: OTPEntryInterface[]
-  ) {
-    if (siteName.length < 2) {
-      return false;
-    }
-
-    const matched = [];
-
-    for (const entry of entries) {
-      if (this.isMatchedEntry(siteName, entry)) {
-        matched.push(entry.hash);
-      }
-    }
-
-    return matched;
-  }
-
-  private isMatchedEntry(
-    siteName: Array<string | null>,
-    entry: OTPEntryInterface
-  ) {
-    if (!entry.issuer) {
-      return false;
-    }
-
-    const issuerHostMatches = entry.issuer.split("::");
-    const issuer = issuerHostMatches[0]
-      .replace(/[^0-9a-z]/gi, "")
-      .toLowerCase();
-
-    if (!issuer) {
-      return false;
-    }
-
-    const siteTitle = siteName[0] || "";
-    const siteNameFromHost = siteName[1] || "";
-    const siteHost = siteName[2] || "";
-
-    if (issuerHostMatches.length > 1) {
-      if (siteHost && siteHost.indexOf(issuerHostMatches[1]) !== -1) {
-        return true;
-      }
-    }
-    // site title should be more detailed
-    // so we use siteTitle.indexOf(issuer)
-    if (siteTitle && siteTitle.indexOf(issuer) !== -1) {
-      return true;
-    }
-
-    if (siteNameFromHost && issuer.indexOf(siteNameFromHost) !== -1) {
-      return true;
-    }
-
-    return false;
   }
 }
